rename(advanced): 포인트 계산 로직 파일 이름 변경
