+++ src/advanced/usecase/calculateBonusPoints.ts
